+++ src/main/java/groom/him/domain/point/models/dto/response/PointResponse.java
@@ -1,7 +1,7 @@
 package groom.him.domain.point.models.dto.response;
 
 public record PointResponse (
-    Long point,
-    Long tbdPoint,
-    Long tbePoint
+    Integer point,
+    Integer tbdPoint,
+    Integer tbePoint
 ){ }
